Modified Email Functionality
